src/requirements.txt
annotated-types==0.7.0
anydi==0.26.5
asgiref==3.8.1
Django==5.0.6
django-ninja==1.1.0
pydantic==2.7.3
pydantic_core==2.18.4
sqlparse==0.5.0
typing_extensions==4.12.2
